requirements.txt
python-telegram-bot>=22.1
psutil>=5.9.0
python-dotenv>=1.0.0